refactor(frontend): unify use nft composable
