+++ frontend/app/src/modules/balances/non-fungible/composables/use-nft-data.ts
@@ -7,27 +7,40 @@ import type { NonFungibleBalance, NonFungibleBalancesRequestPayload } from '@/ty
 import { usePaginationFilters } from '@/composables/use-pagination-filter';
 import { useNftBalances } from '@/modules/balances/nft/use-nft-balances';
 import { TableId, useRememberTableSorting } from '@/modules/table/use-remember-table-sorting';
+import { useFrontendSettingsStore } from '@/store/settings/frontend';
 import { useGeneralSettingsStore } from '@/store/settings/general';
+import { useStatisticsStore } from '@/store/statistics';
 import { useStatusStore } from '@/store/status';
+import { DashboardTableType } from '@/types/settings/frontend-settings';
 import { Section } from '@/types/status';
+import { TableColumn } from '@/types/table-column';
+import { calculatePercentage } from '@/utils/calculation';
 import { getCollectionData } from '@/utils/collection';
+
+interface UseNftDataOptions {
+  dashboard?: boolean;
+}
 
 interface UseNftDataReturn {
   balances: Ref<Collection<NonFungibleBalance>>;
+  cols: ComputedRef<DataTableColumn<NonFungibleBalance>[]>;
   currencySymbol: ComputedRef<string>;
   data: ComputedRef<NonFungibleBalance[]>;
   dataLoading: Ref<boolean>;
   fetchData: () => Promise<void>;
   ignoredAssetsHandling: Ref<IgnoredAssetsHandlingType>;
   pagination: ComputedRef<TablePaginationData>;
+  percentageOfCurrentGroup: (value: BigNumber) => string;
+  percentageOfTotalNetValue: (value: BigNumber) => string;
   refreshNonFungibleBalances: (ignoreCache?: boolean) => Promise<void>;
   sectionLoading: ComputedRef<boolean>;
   sort: ComputedRef<DataTableSortData<NonFungibleBalance>>;
-  cols: ComputedRef<DataTableColumn<NonFungibleBalance>[]>;
   totalValue: ComputedRef<BigNumber | undefined | null>;
 }
 
-export function useNftData(): UseNftDataReturn {
+export function useNftData(options: UseNftDataOptions = {}): UseNftDataReturn {
+  const { dashboard = false } = options;
+
   const { fetchNonFungibleBalances, refreshNonFungibleBalances } = useNftBalances();
   const { currencySymbol } = storeToRefs(useGeneralSettingsStore());
   const { t } = useI18n({ useScope: 'global' });
@@ -37,38 +50,6 @@ export function useNftData(): UseNftDataReturn {
   const extraParams = computed(() => ({
     ignoredAssetsHandling: get(ignoredAssetsHandling),
   }));
-
-  const cols = computed<DataTableColumn<NonFungibleBalance>[]>(() => [{
-    cellClass: 'text-no-wrap',
-    key: 'name',
-    label: t('common.name'),
-    sortable: true,
-  }, {
-    align: 'center',
-    key: 'ignored',
-    label: t('non_fungible_balances.ignore'),
-  }, {
-    align: 'end',
-    class: 'text-no-wrap',
-    key: 'priceInAsset',
-    label: t('non_fungible_balances.column.price_in_asset'),
-    width: '75%',
-  }, {
-    align: 'end',
-    class: 'text-no-wrap',
-    key: 'price',
-    label: t('common.price_in_symbol', { symbol: get(currencySymbol) }),
-    sortable: true,
-  }, {
-    class: 'text-no-wrap',
-    key: 'manuallyInput',
-    label: t('non_fungible_balances.column.custom_price'),
-  }, {
-    align: 'center',
-    key: 'actions',
-    label: t('common.actions_text'),
-    width: '50',
-  }]);
 
   const { isLoading: isSectionLoading } = useStatusStore();
   const sectionLoading = isSectionLoading(Section.NON_FUNGIBLE_BALANCES);
@@ -89,21 +70,125 @@ export function useNftData(): UseNftDataReturn {
       direction: 'desc',
     }],
     extraParams,
-    history: 'router',
-    onUpdateFilters(query) {
-      set(ignoredAssetsHandling, query.ignoredAssetsHandling || 'exclude');
-    },
+    history: dashboard ? 'external' : 'router',
+    ...(!dashboard && {
+      onUpdateFilters(query): void {
+        set(ignoredAssetsHandling, query.ignoredAssetsHandling || 'exclude');
+      },
+    }),
+  });
+
+  const { data, totalValue } = getCollectionData(balances);
+
+  // Watch ignoredAssetsHandling changes and reset to page 1 (only for non-dashboard)
+  if (!dashboard) {
+    watch(ignoredAssetsHandling, () => {
+      setPage(1);
+    });
+  }
+
+  // Dashboard-specific: percentage calculations
+  const statistics = useStatisticsStore();
+  const { totalNetWorth } = storeToRefs(statistics);
+  const { dashboardTablesVisibleColumns } = storeToRefs(useFrontendSettingsStore());
+
+  function percentageOfTotalNetValue(value: BigNumber): string {
+    return calculatePercentage(value, get(totalNetWorth));
+  }
+
+  function percentageOfCurrentGroup(value: BigNumber): string {
+    return calculatePercentage(value, get(totalValue) as BigNumber);
+  }
+
+  // Columns configuration
+  const cols = computed<DataTableColumn<NonFungibleBalance>[]>(() => {
+    if (dashboard) {
+      const group = DashboardTableType.NFT;
+      const visibleColumns = get(dashboardTablesVisibleColumns)[group];
+
+      const headers: DataTableColumn<NonFungibleBalance>[] = [{
+        cellClass: 'py-0',
+        class: 'text-no-wrap w-full',
+        key: 'name',
+        label: t('common.name'),
+        sortable: true,
+      }, {
+        align: 'end',
+        cellClass: 'py-0',
+        class: 'text-no-wrap',
+        key: 'priceInAsset',
+        label: t('nft_balance_table.column.price_in_asset'),
+      }, {
+        align: 'end',
+        cellClass: 'py-0',
+        class: 'text-no-wrap',
+        key: 'price',
+        label: t('common.price_in_symbol', {
+          symbol: get(currencySymbol),
+        }),
+        sortable: true,
+      }];
+
+      if (visibleColumns.includes(TableColumn.PERCENTAGE_OF_TOTAL_NET_VALUE)) {
+        headers.push({
+          align: 'end',
+          cellClass: 'py-0',
+          class: 'text-no-wrap',
+          key: 'percentageOfTotalNetValue',
+          label: t('nft_balance_table.column.percentage'),
+        });
+      }
+
+      if (visibleColumns.includes(TableColumn.PERCENTAGE_OF_TOTAL_CURRENT_GROUP)) {
+        headers.push({
+          align: 'end',
+          cellClass: 'py-0',
+          class: 'text-no-wrap',
+          key: 'percentageOfTotalCurrentGroup',
+          label: t('dashboard_asset_table.headers.percentage_of_total_current_group', {
+            group,
+          }),
+        });
+      }
+
+      return headers;
+    }
+
+    // Non-dashboard columns
+    return [{
+      cellClass: 'text-no-wrap',
+      key: 'name',
+      label: t('common.name'),
+      sortable: true,
+    }, {
+      align: 'center',
+      key: 'ignored',
+      label: t('non_fungible_balances.ignore'),
+    }, {
+      align: 'end',
+      class: 'text-no-wrap',
+      key: 'priceInAsset',
+      label: t('non_fungible_balances.column.price_in_asset'),
+      width: '75%',
+    }, {
+      align: 'end',
+      class: 'text-no-wrap',
+      key: 'price',
+      label: t('common.price_in_symbol', { symbol: get(currencySymbol) }),
+      sortable: true,
+    }, {
+      class: 'text-no-wrap',
+      key: 'manuallyInput',
+      label: t('non_fungible_balances.column.custom_price'),
+    }, {
+      align: 'center',
+      key: 'actions',
+      label: t('common.actions_text'),
+      width: '50',
+    }];
   });
 
   useRememberTableSorting<NonFungibleBalance>(TableId.NON_FUNGIBLE_BALANCES, sort, cols);
-
-  // Extract collection data
-  const { data, totalValue } = getCollectionData(balances);
-
-  // Watch ignoredAssetsHandling changes and reset to page 1
-  watch(ignoredAssetsHandling, () => {
-    setPage(1);
-  });
 
   return {
     balances,
@@ -114,6 +199,8 @@ export function useNftData(): UseNftDataReturn {
     fetchData,
     ignoredAssetsHandling,
     pagination,
+    percentageOfCurrentGroup,
+    percentageOfTotalNetValue,
     refreshNonFungibleBalances,
     sectionLoading,
     sort,
